refactor: 오타 수정 및 변수명 수정
