+++ order-service/src/main/java/com/firstlogistics/orderservice/application/port/dto/CompanyResponse.java
@@ -1,7 +1,10 @@
 package com.firstlogistics.orderservice.application.port.dto;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
 import java.util.UUID;
 
+@JsonIgnoreProperties(ignoreUnknown = true)
 public record CompanyResponse(
         UUID id,
         UUID hubId,
