+++ order-service/src/main/java/com/firstlogistics/orderservice/infrastructure/feign/HubClient.java
@@ -1,6 +1,6 @@
 package com.firstlogistics.orderservice.infrastructure.feign;
 
-import com.firstlogistics.orderservice.infrastructure.feign.dto.HubManagerResponse;
+import com.firstlogistics.orderservice.application.port.dto.HubManagerResponse;
 import common.response.ApiResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
