+++ order-service/src/main/java/com/firstlogistics/orderservice/infrastructure/feign/adapter/CompanyFeignAdapter.java
@@ -21,9 +21,9 @@ public class CompanyFeignAdapter implements CompanyPort {
     private final CompanyClient companyClient;
 
     @Override
-    public CompanyResponse getCompanyById(UUID managerId) {
+    public CompanyResponse getCompanyById(UUID companyId) {
         try {
-            ApiResponse<CompanyResponse> response = companyClient.getCompanyById(managerId);
+            ApiResponse<CompanyResponse> response = companyClient.getCompanyById(companyId);
 
             if (response != null && response.getStatus().is2xxSuccessful() && response.getData() != null) {
                 return response.getData();
@@ -32,7 +32,7 @@ public class CompanyFeignAdapter implements CompanyPort {
             throw new OrderException(OrderErrorCode.COMPANY_NOT_FOUND);
 
         } catch (FeignException.NotFound e) {
-            log.warn("업체 관리자 정보를 찾을 수 없습니다. managerId: {}", managerId);
+            log.warn("업체 관리자 정보를 찾을 수 없습니다. companyId: {}", companyId);
             throw new OrderException(OrderErrorCode.COMPANY_NOT_FOUND);
         } catch (Exception e) {
             log.error("업체 서비스 통신 중 오류 발생: {}", e.getMessage());
